// src/server/types.ts
/**
 * 境界表
 */
export interface LevelListType {
  UID: string
  autograph: string
  lifeName: string
  levelName: string
  power: number
  user_avatar: string
}
/**
 * 杀神表
 */
export interface KillListType {
  UID: string
  autograph: string
  lifeName: string
  prestige: number
  power: number
  user_avatar: string
}